Work to address extreme