@danielballan 2021-07-05
Tested interactively on filled and unfilled data at CSX.

Timing for reading a table with dimensions 2 rows x 269 columns is greatly improved by about 5X. The absolute time (~400ms) is acceptable.

Before (at commit b13dd52e390a837c57bc149e38d6eabc6da4adf5):

```py
In [4]: %timeit h.table("baseline")
2.22 s ± 31.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After (at commit cfeedbc5488ca2355d3e7183cebe3c860bfdc525):

```py
In [4]: %timeit h.table("baseline")
406 ms ± 2.31 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```